# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Payments API",
    "version": "1.0.0",
    "description": "API for managing Payments."
  },
  "servers": [
    {
      "url": "http://localhost:5002/api",
      "description": "Local server"
    }
  ],
  "paths": {
    "/payments": {
      "get": {
        "summary": "Get all Payments",
        "operationId": "getAllPayments",
        "tags": ["Payments"],
        "responses": {
          "200": {
            "description": "List of Payments",
            "content": {
              "application/json": {
                "example": [
                  {
                    "paymentID": "1a2b3c4d-5678-90ab-cdef-1234567890ab",
                    "totalValue": 100.50,
                    "paymentType": "Credit Card",
                    "paymentStatusID": "1",
                    "ticketID": "9876543210",
                    "createdAt": "2024-11-18T12:00:00Z"
                  }
                ]
              }
            }
          },
          "404": {
            "description": "No payments found"
          },
          "500": {
            "description": "Internal error while fetching payments"
          }
        }
      },
      "post": {
        "summary": "Create a new Payment",
        "operationId": "createPayment",
        "tags": ["Payments"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "totalValue": 100.50,
                "paymentType": "Credit Card",
                "paymentStatusID": "1",
                "ticketID": "9876543210"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment created successfully",
            "content": {
              "application/json": {
                "example": {
                  "paymentID": "1a2b3c4d-5678-90ab-cdef-1234567890ab",
                  "totalValue": 100.50,
                  "paymentType": "Credit Card",
                  "paymentStatusID": "1",
                  "ticketID": "9876543210",
                  "createdAt": "2024-11-18T12:00:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields"
          },
          "500": {
            "description": "Error creating Payment"
          }
        }
      }
    },
    "/payments/{id}": {
      "get": {
        "summary": "Get a Payment by ID",
        "operationId": "getPaymentById",
        "tags": ["Payments"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1a2b3c4d-5678-90ab-cdef-1234567890ab"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment retrieved successfully",
            "content": {
              "application/json": {
                "example": {
                  "paymentID": "1a2b3c4d-5678-90ab-cdef-1234567890ab",
                  "totalValue": 100.50,
                  "paymentType": "Credit Card",
                  "paymentStatusID": "1",
                  "ticketID": "9876543210",
                  "createdAt": "2024-11-18T12:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "Payment not found"
          },
          "500": {
            "description": "Error fetching Payment by ID"
          }
        }
      },
      "put": {
        "summary": "Update a Payment",
        "operationId": "updatePayment",
        "tags": ["Payments"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1a2b3c4d-5678-90ab-cdef-1234567890ab"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "totalValue": 120.75,
                "paymentType": "Debit Card",
                "paymentStatusID": "2",
                "ticketID": "9876543210"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment updated successfully",
            "content": {
              "application/json": {
                "example": {
                  "paymentID": "1a2b3c4d-5678-90ab-cdef-1234567890ab",
                  "totalValue": 120.75,
                  "paymentType": "Debit Card",
                  "paymentStatusID": "2",
                  "ticketID": "9876543210",
                  "createdAt": "2024-11-18T12:00:00Z"
                }
              }
            }
          },
          "404": {
            "description": "Payment not found"
          },
          "500": {
            "description": "Error updating Payment"
          }
        }
      },
      "delete": {
        "summary": "Delete a Payment",
        "operationId": "deletePayment",
        "tags": ["Payments"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1a2b3c4d-5678-90ab-cdef-1234567890ab"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Payment deleted successfully"
          },
          "404": {
            "description": "Payment not found"
          },
          "500": {
            "description": "Error deleting Payment"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Payment": {
        "type": "object",
        "properties": {
          "paymentID": {
            "type": "string",
            "example": "1a2b3c4d-5678-90ab-cdef-1234567890ab"
          },
          "totalValue": {
            "type": "number",
            "example": 100.50
          },
          "paymentType": {
            "type": "string",
            "example": "Credit Card"
          },
          "paymentStatusID": {
            "type": "string",
            "example": "1"
          },
          "ticketID": {
            "type": "string",
            "example": "9876543210"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-11-18T12:00:00Z"
          }
        }
      }
    }
  }
}